+++ splitcal/splitcalCluster.cxx
@@ -22,24 +22,9 @@
 splitcalCluster::splitcalCluster()
 {
 }
-// -----   constructor from splitcalHit   ------------------------------------------
-splitcalCluster::splitcalCluster(splitcalHit* h)
-{
-  _vectorOfHits.push_back(h);
-  _hitWeights.push_back(1.0);  // Default weight is 1.0
-
-  if ( _vectorOfHits.size() == 1){ //first element added to the cluster
-    SetStartPoint(h);
-    SetEndPoint(h);
-  } else {
-    if (_start.Z() > h->GetZ()) SetStartPoint(h); // start point is the hit with the smalllest z
-    if (_end.Z() < h->GetZ()) SetEndPoint(h); // end point is the hit with the largest z
-  }
-
-}
 
 
-void splitcalCluster::ComputeEtaPhiE()
+void splitcalCluster::ComputeEtaPhiE(const std::vector<splitcalHit>& hits)
 {
 
 
@@ -56,29 +41,29 @@ void splitcalCluster::ComputeEtaPhiE()
 
   // loop over hits to compute cluster energy sum and to compute the coordinates weighted average per layer
   double energy = 0.;
-  for (size_t i = 0; i < _vectorOfHits.size(); ++i){
-    auto hit = _vectorOfHits[i];
-    double hitEnergy = hit->GetEnergy() * _hitWeights[i];  // Use weight from cluster
+  for (size_t i = 0; i < _hitIndices.size(); ++i){
+    const auto& hit = hits[_hitIndices[i]];
+    double hitEnergy = hit.GetEnergy() * _hitWeights[i];  // Use weight from cluster
     energy += hitEnergy;
-    int layer = hit->GetLayerNumber();
+    int layer = hit.GetLayerNumber();
     // hits from high precision layers give both x and y coordinates --> use if-if instead of if-else
-    if (hit->IsX()){
+    if (hit.IsX()){
       if (mapLayerWeigthedX.count(layer)==0) { //if key is not yet in map, initialise element to 0
 	mapLayerWeigthedX[layer] = 0.;
 	mapLayerSumWeigthsX[layer] = 0.;
       }
-      mapLayerWeigthedX[layer] += hit->GetX() * hitEnergy;
+      mapLayerWeigthedX[layer] += hit.GetX() * hitEnergy;
       mapLayerSumWeigthsX[layer] += hitEnergy;
-      mapLayerZ1[layer] = hit->GetZ();
+      mapLayerZ1[layer] = hit.GetZ();
     }
-    if (hit->IsY()){
+    if (hit.IsY()){
       if (mapLayerWeigthedY.count(layer)==0) { //if key is not yet in map, initialise element to 0
 	mapLayerWeigthedY[layer] = 0.;
 	mapLayerSumWeigthsY[layer] = 0.;
       }
-      mapLayerWeigthedY[layer] += hit->GetY() * hitEnergy;
+      mapLayerWeigthedY[layer] += hit.GetY() * hitEnergy;
       mapLayerSumWeigthsY[layer] += hitEnergy;
-      mapLayerZ2[layer] = hit->GetZ();
+      mapLayerZ2[layer] = hit.GetZ();
     }
   }//end loop on hit
 
@@ -113,9 +98,10 @@ void splitcalCluster::ComputeEtaPhiE()
 
   SetEndPoint(maxX, maxY, maxZ);
 
-  // get direction vector from end-strat vector difference
-  TVector3 direction;
-  direction = _end - _start;
+  // get direction vector from end-start vector difference
+  TVector3 start(_start[0], _start[1], _start[2]);
+  TVector3 end(_end[0], _end[1], _end[2]);
+  TVector3 direction = end - start;
   double eta = direction.Eta();
   double phi = direction.Phi();
   SetEtaPhiE(eta, phi, energy);
@@ -209,15 +195,6 @@ regression splitcalCluster::LinearRegression(std::vector<double >& x, std::vecto
 }
 
 
-void splitcalCluster::SetStartPoint(splitcalHit*& h) {
-  SetStartPoint(h->GetX(),h->GetY(),h->GetZ());
-}
-
-void  splitcalCluster::SetEndPoint(splitcalHit*& h) {
-  SetEndPoint(h->GetX(),h->GetY(),h->GetZ());
-}
-
-
 // -------------------------------------------------------------------------
 
 // -----   Destructor   ----------------------------------------------------
@@ -234,13 +211,13 @@ void splitcalCluster::Print() const
 	   << _phi << " ,  "
 	   << _energy << std::endl;
   std::cout<< "    start(x,y,z) = "
-	   << _start.X() << " ,  "
-	   << _start.Y() << " ,  "
-	   << _start.Z() << std::endl;
+	   << _start[0] << " ,  "
+	   << _start[1] << " ,  "
+	   << _start[2] << std::endl;
   std::cout<< "    end(x,y,z) = "
-	   << _end.X() << " ,  "
-	   << _end.Y() << " ,  "
-	   << _end.Z() << std::endl;
+	   << _end[0] << " ,  "
+	   << _end[1] << " ,  "
+	   << _end[2] << std::endl;
    std::cout<< "------- " <<std::endl;
 }
 // -------------------------------------------------------------------------